extract board print functions to separate module
